Update VFX
*Adicionado um local de recuperação de vida
*Efeito visual de bolha na área
*Adicionado particulas nas coletas (moedas, armas e pacotes de vida)
*Adicionado particulas (de bolhas) ao utilizar o pacote de vida
*Adicionado particulas ao correr e ao aterrisar com o personagem
*Adicionado rastros nos projeteis

+++ Assets/Scripts/DestructableItemBase/DestructableItemBase.cs
@@ -47,15 +47,15 @@ public class DestructableItemBase : MonoBehaviour, IDamageable
         }
     }
 
-    public void Damage(int damage)
+    public void Damage(float damage)
     {
     }
 
-    public void Damage(int damage, Transform pos)
+    public void Damage(float damage, Transform pos)
     {
     }
 
-    public void Damage(int damage, Transform pos, bool recoil, bool constant)
+    public void Damage(float damage, Transform pos, bool recoil, bool constant)
     {
         if(0 <= savedCoins)
         {
