+++ Assets/Scripts/Interfaces/IDamageable.cs
@@ -4,11 +4,12 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    void Damage(int damage);
+    void Damage(float damage);
 
 
-    void Damage(int damage, Transform pos);
+    void Damage(float damage, Transform pos);
 
 
-    void Damage(int damage, Transform pos, bool recoil, bool constant);
+    void Damage(float damage, Transform pos, bool recoil, bool constant);
+
 }
